refac: Apply new refactor changes to some basic vuln detector tests #2462
