+++ deps/wazuh_testing/wazuh_testing/modules/vulnerability_detector/__init__.py
@@ -1,6 +1,7 @@
 import os
 
 from wazuh_testing.tools import WAZUH_PATH
+from wazuh_testing.db_interface import CVE_DB_PATH
 
 
 VULN_DETECTOR_GLOBAL_TIMEOUT = 20
@@ -11,7 +12,6 @@ DEBIAN_IMPORT_FEED_TIMEOUT = 50
 DOWNLOAD_TIMEOUT = 140
 
 DB_PATH = os.path.join(WAZUH_PATH, 'queue', 'db')
-CVE_DB_PATH = os.path.join(WAZUH_PATH, 'queue', 'vulnerabilities', 'cve.db')
 MSU_PATH = os.path.join(WAZUH_PATH, 'queue', 'vulnerabilities', 'dictionaries', 'msu.json.gz')
 CPE_HELPER_PATH = os.path.join(WAZUH_PATH, 'queue', 'vulnerabilities', 'dictionaries', 'cpe_helper.json')
 DEFAULT_PACKAGE_NAME = "wazuhintegrationpackage"
@@ -109,15 +109,11 @@ CUSTOM_MSU = 'custom_msu.json'
 CUSTOM_CPE_HELPER = 'custom_cpe_helper.json'
 VULNERABILITIES = 'vulnerabilities.json'
 
-NVD_TABLES = [
-    {'name': 'NVD_REFERENCE', 'path': CVE_DB_PATH},
-    {'name': 'NVD_METRIC_CVSS', 'path': CVE_DB_PATH},
-    {'name': 'NVD_METADATA', 'path': CVE_DB_PATH},
-    {'name': 'NVD_CVE_MATCH', 'path': CVE_DB_PATH},
-    {'name': 'NVD_CVE_CONFIGURATION', 'path': CVE_DB_PATH},
-    {'name': 'NVD_CVE', 'path': CVE_DB_PATH},
-    {'name': 'NVD_CPE', 'path': CVE_DB_PATH},
-]
+FEED_TABLES = ['vulnerabilities', 'vulnerabilities_info', 'references_info', 'bugzilla_references_info',
+               'advisories_info', 'metadata', 'variables']
+AGENT_SYS_TABLES = ['sys_programs', 'sys_hotfixes', 'sys_osinfo']
+NVD_TABLES = ['NVD_REFERENCE', 'NVD_METRIC_CVSS', 'NVD_METADATA', 'NVD_CVE_MATCH', 'NVD_CVE_MATCH',
+              'NVD_CVE_CONFIGURATION', 'NVD_CVE', 'NVD_CPE']
 
 REDHAT_KEY_FIELDS_FEEDS = ['CVE', 'bugzilla_description', 'affected_packages']
