+++ deps/wazuh_testing/wazuh_testing/modules/vulnerability_detector/event_monitor.py
@@ -4,7 +4,7 @@ from wazuh_testing.modules import vulnerability_detector as vd
 from wazuh_testing.db_interface.cve_db import get_num_vulnerabilities
 
 
-def _make_vuln_callback(pattern, prefix=vd.VULNERABILITY_DETECTOR_PREFIX):
+def make_vuln_callback(pattern, prefix=vd.VULNERABILITY_DETECTOR_PREFIX):
     """Create a callback function from a text pattern.
 
     It already contains the vulnerability-detector prefix.
@@ -62,7 +62,7 @@ def check_vuln_detector_event(wazuh_log_monitor, callback, error_message='', upd
         prefix (str): log pattern regex
     """
     wazuh_log_monitor.start(timeout=timeout, update_position=update_position,
-                            callback=_make_vuln_callback(callback, prefix), error_message=error_message)
+                            callback=make_vuln_callback(callback, prefix), error_message=error_message)
 
 
 def check_vulnerabilities_number(expected_number):
