+++ tests/integration/test_vulnerability_detector/conftest.py
@@ -12,7 +12,10 @@ from wazuh_testing.tools import LOG_FILE_PATH
 from wazuh_testing.tools.file import truncate_file
 from wazuh_testing.tools.monitoring import FileMonitor
 from wazuh_testing.tools.services import control_service
-from wazuh_testing import vulnerability_detector as vd
+from wazuh_testing.modules.vulnerability_detector.utils import clean_vd_tables
+from wazuh_testing.mocking import create_mocked_agent, delete_mocked_agent
+from wazuh_testing import db_interface as dbi
+from wazuh_testing.modules import vulnerability_detector as vd
 
 
 @pytest.fixture(scope='module')
@@ -31,11 +34,11 @@ def restart_modulesd(get_configuration, request):
 @pytest.fixture(scope='module')
 def clean_vuln_tables(request):
     """Clean vulnerabilities tables """
-    vd.clean_vd_tables()
+    clean_vd_tables()
 
     yield
 
-    vd.clean_vd_tables()
+    clean_vd_tables()
 
 
 @pytest.fixture
@@ -55,11 +58,11 @@ def mock_agent():
     """
     Fixture to create a mocked agent in wazuh databases
     """
-    agent_id = vd.create_mocked_agent(name="mocked_agent")
+    agent_id = create_mocked_agent(name="mocked_agent")
 
     yield agent_id
 
-    vd.delete_mocked_agent(agent_id)
+    delete_mocked_agent(agent_id)
 
 
 @pytest.fixture(scope='module')
@@ -67,11 +70,11 @@ def mock_agent_module():
     """
     Fixture to create a mocked agent in wazuh databases
     """
-    agent_id = vd.create_mocked_agent(name="mocked_agent")
+    agent_id = create_mocked_agent(name="mocked_agent")
 
     yield agent_id
 
-    vd.delete_mocked_agent(agent_id)
+    delete_mocked_agent(agent_id)
 
 
 @pytest.fixture(scope='module')
@@ -81,7 +84,7 @@ def check_cve_db():
     """
     def cve_tables_created():
         query_string = "SELECT count(*) FROM sqlite_master WHERE type='table'"
-        query_result = int(vd.get_query_result(vd.CVE_DB_PATH, query_string)[0])
+        query_result = int(dbi.get_query_result(dbi.CVE_DB_PATH, query_string)[0])
         return query_result == vd.CVE_NUM_TABLES
 
     db_status_ok = False
@@ -104,7 +107,7 @@ def after_execution_clean_vuln_tables_function(request):
 
     yield
 
-    vd.clean_vd_tables()
+    clean_vd_tables()
 
 
 @pytest.fixture(scope='function')
