+++ tests/integration/test_vulnerability_detector/test_general_settings/test_general_settings_enabled.py
@@ -57,14 +57,15 @@ tags:
     - vulnerability_detector
 '''
 import os
-
 import pytest
+
 from wazuh_testing.tools import LOG_FILE_PATH
 from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
 from wazuh_testing.tools.monitoring import FileMonitor
-from wazuh_testing.vulnerability_detector import callback_detect_vulnerability_detector_disabled, \
-    callback_detect_vulnerability_detector_enabled, \
-    VULN_DETECTOR_EXTENDED_GLOBAL_TIMEOUT
+from wazuh_testing.modules.vulnerability_detector import VULN_DETECTOR_EXTENDED_GLOBAL_TIMEOUT
+from wazuh_testing.modules.vulnerability_detector.event_monitor import callback_detect_vulnerability_detector_enabled, \
+                                                                       callback_detect_vulnerability_detector_disabled
+
 
 # Marks
 pytestmark = [pytest.mark.server, pytest.mark.tier(level=0)]
