+++ tests/integration/test_vulnerability_detector/test_general_settings/test_general_settings_interval.py
@@ -55,14 +55,15 @@ tags:
     - vulnerability_detector
 '''
 import os
-
 import pytest
+
 from wazuh_testing.tools import LOG_FILE_PATH
 from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
 from wazuh_testing.tools.monitoring import FileMonitor
 from wazuh_testing.tools.time import time_to_human_readable, unit_to_seconds
-from wazuh_testing.vulnerability_detector import callback_detect_vulnerability_scan_sleeping, \
-    VULN_DETECTOR_EXTENDED_GLOBAL_TIMEOUT
+from wazuh_testing.modules.vulnerability_detector import VULN_DETECTOR_EXTENDED_GLOBAL_TIMEOUT
+from wazuh_testing.modules.vulnerability_detector.event_monitor import callback_detect_vulnerability_scan_sleeping
+
 
 # Marks
 pytestmark = [pytest.mark.server, pytest.mark.tier(level=0)]
