+++ tests/integration/test_vulnerability_detector/test_general_settings/test_general_settings_retry_interval.py
@@ -52,17 +52,19 @@ tags:
     - settings
 '''
 import os
-import time
 import re
 import math
 import pytest
-import wazuh_testing.vulnerability_detector as vd
-
 from datetime import datetime
+
 from wazuh_testing.tools import LOG_FILE_PATH
 from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
 from wazuh_testing.tools.monitoring import FileMonitor
 from wazuh_testing.tools.time import time_to_seconds
+from wazuh_testing.modules import vulnerability_detector as vd
+from wazuh_testing.modules.vulnerability_detector.event_monitor import make_vuln_callback
+from wazuh_testing.db_interface import agent_db
+from wazuh_testing.modules.vulnerability_detector.utils import clean_vd_tables
 
 # Marks
 pytestmark = [pytest.mark.server, pytest.mark.tier(level=0)]
@@ -107,14 +109,14 @@ def mock_system(mock_agent):
     """
     It mocks an agent with syscollector packages in a not synced state.
     """
-    vd.insert_osinfo(agent=mock_agent)
+    agent_db.insert_os_info(agent_id=mock_agent)
 
     # Force not synced state for syscollector packages
-    vd.update_sync_info(agent=mock_agent, last_attempt=0, last_completion=0)
+    agent_db.update_sync_info(agent_id=mock_agent, last_attempt=0, last_completion=0)
 
     yield mock_agent
 
-    vd.clean_vd_tables(mock_agent)
+    clean_vd_tables(mock_agent)
 
 
 def test_retry_interval(get_configuration, configure_environment, restart_modulesd, mock_system):
@@ -153,7 +155,7 @@ def test_retry_interval(get_configuration, configure_environment, restart_module
     retry_interval = time_to_seconds(get_configuration['metadata']['retry_interval'])
 
     def get_timestamp(log_line):
-        callback_log_line = vd.make_vuln_callback(pattern=log_line, prefix='.*')
+        callback_log_line = make_vuln_callback(pattern=log_line, prefix='.*')
         wazuh_log_monitor.start(timeout=vd.VULN_DETECTOR_SCAN_TIMEOUT,
                                 callback=callback_log_line,
                                 error_message='Log line not found.')
