+++ tests/integration/test_vulnerability_detector/test_general_settings/test_general_settings_run_on_start.py
@@ -53,13 +53,14 @@ tags:
 '''
 
 import os
-
 import pytest
-import wazuh_testing.vulnerability_detector as vd
+
 from wazuh_testing.tools import LOG_FILE_PATH
 from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
 from wazuh_testing.tools.monitoring import FileMonitor
-from wazuh_testing.vulnerability_detector import make_vuln_callback, VULN_DETECTOR_EXTENDED_GLOBAL_TIMEOUT
+from wazuh_testing.modules.vulnerability_detector.event_monitor import make_vuln_callback
+from wazuh_testing.modules.vulnerability_detector import VULN_DETECTOR_EXTENDED_GLOBAL_TIMEOUT, CUSTOM_NVD_FEED
+
 
 # Marks
 pytestmark = [pytest.mark.server, pytest.mark.tier(level=0)]
@@ -69,7 +70,7 @@ current_test_path = os.path.dirname(os.path.realpath(__file__))
 test_data_path = os.path.join(current_test_path, 'data')
 test_feeds_path = os.path.join(current_test_path, '..', 'data')
 configurations_path = os.path.join(test_data_path, 'wazuh_run_on_start.yaml')
-custom_nvd_feed_path = os.path.join(test_feeds_path, 'feeds', vd.CUSTOM_NVD_FEED)
+custom_nvd_feed_path = os.path.join(test_feeds_path, 'feeds', CUSTOM_NVD_FEED)
 
 wazuh_log_monitor = FileMonitor(LOG_FILE_PATH)
 scan_timeout = 60
